html/semantics/sections/the-body-element/sections-body-element-events-content.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Body Element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/master/single-page.html#the-body-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body 	onbeforeunload="myReaction.push('beforeunload')" 
	onblur="myReaction.push('blur')" 
	onerror="myReaction.push('error')" 
	onfocus="myReaction.push('focus')" 
	onfullscreenchange="myReaction.push('fullscreenchange')" 
	onfullscreenerror="myReaction.push('fullscreenerror')" 
	onhashchange="myReaction.push('hashchange')" 
	onload="myReaction.push('load')" 
	onmessage="myReaction.push('message')"  
	onpagehide="myReaction.push('pagehide')" 
	onpageshow="myReaction.push('pageshow')" 
	onpopstate="myReaction.push('popstate')" 
	onresize="myReaction.push('resize')" 
	onscroll="myReaction.push('scroll')" 
	onstorage="myReaction.push('storage')" 
	onunload="myReaction.push('unload')">

    <h1>Description</h1>
    <p>This test validates that all automatable (16 out of 20) body element content attribute event handlers react to relevant events.</p>

    <div id="log"></div>

    <script>
        "use strict";

	// Verify that all the content attributes that allow body to react to certain events actually do react to those events

	var myReaction = [], num = 0, strEvent="";
	var myEvents = [["error", true, false],["hashchange", true, false], ["resize", true, false], ["scroll", true, false], ["beforeunload", false, true], ["blur", false, false], ["focus", false, false], ["load", false, false], ["unload", false, false], ["fullscreenchange", false, false], ["fullscreenerror", false, false], ["pagehide", false, false], ["pageshow", false, false], ["popstate", false, false], ["storage", false, false]];

	for (num = 0; num < myEvents.length; num++) {

		strEvent = myEvents[num][0];

		test(function() {

			var myEvent = {};

			// double-register event with event listener to make sure something happened for content attribute to react to
			document.body.addEventListener(strEvent, function(event) { myReaction.push(strEvent + "-listener"); }, false);

			// Create and cause the event
			myEvent = document.createEvent("Event");
                	myEvent.initEvent(strEvent, myEvents[num][1], myEvents[num][2]);
			document.body.dispatchEvent(myEvent);
 
			// Check that event fired by looking for event listener results
			assert_true(myReaction.indexOf(strEvent + "-listener") > -1, strEvent + " DID NOT FIRE: ");			

			// Check that content attribute's event handler fired
			assert_true(myReaction.indexOf(strEvent) > -1, "on" + strEvent + " content attribute failed: ");
             	
            	}, "Body's on" + strEvent + " content attribute responds to event.");

	}

	// Now handle the message event asynchronously
	var msgTest = async_test("Body's onmessage content attribute responds to event.");

	msgTest.step(function () {
		// add this event listener as a triple-check that event fires and is caught by this test
		window.addEventListener("message", function(event) { myReaction.push("message-WINDOW"); }, false);
 	});

	msgTest.step(function () {
		window.postMessage("message_here", "*");
    	});

	msgTest.step(function () {
		window.onmessage = msgTest.step_func(function (event) {
			assert_true(myReaction.indexOf("message-WINDOW") > -1, "MESSAGE EVENT NOT REGISTERED BY WINDOW EV LISTENER: ");        	
			assert_true(myReaction.indexOf("message") > -1, "onmessage content attribute failed: ");
			msgTest.done();
		});
    	});

    </script>

</body>
